Add hadle exceptions for methods patch and delete

+++ src/services/users.py
@@ -35,10 +35,10 @@ class UsersService:
                     "При получении списка пользователей произошла ошибка."
                 ) from e
 
-    async def find_user(self, uuid: UUID):
+    async def find_user(self, user_uuid: UUID):
         async with unit_of_work() as uow:
             try:
-                user = await self.tasks_repo.find(uow.session, uuid)
+                user = await self.tasks_repo.find(uow.session, user_uuid)
                 print(user)
                 if user is None:
                     raise EntityNotFoundException("Пользователь не найден.")
@@ -52,15 +52,32 @@ class UsersService:
 
     async def patch_user(self, user_uuid: UUID, user_update: UserProfilePatch):
         user_update_dict = user_update.model_dump(exclude_defaults=True)
-        async with unit_of_work() as uow:
-            res = await self.tasks_repo.patch(uow.session, user_uuid, user_update_dict)
-            if res == 1:
-                return True
-            return False
+        try:
+            async with unit_of_work() as uow:
+                user = await self.tasks_repo.find(uow.session, user_uuid)
+                if user is None:
+                    raise EntityNotFoundException("Пользователь не найден.")
+                await self.tasks_repo.patch(uow.session, user_uuid, user_update_dict)
+        except EntityNotFoundException:
+            raise
+        except Exception as e:
+            raise BusinessValidationError(
+                "При изменении сведений о пользователе произошла ошибка."
+            ) from e
 
     async def delte_user(self, user_uuid: UUID):
-        async with unit_of_work() as uow:
-            res = await self.tasks_repo.delete(uow.session, user_uuid)
-            if res == 1:
-                return True
-            return False
+        try:
+            async with unit_of_work() as uow:
+                user = await self.tasks_repo.find(uow.session, user_uuid)
+                if user is None:
+                    raise EntityNotFoundException("Пользователь не найден.")
+                res = await self.tasks_repo.delete(uow.session, user_uuid)
+                if res == 1:
+                    return True
+                return False
+        except EntityNotFoundException:
+            raise
+        except Exception as e:
+            raise BusinessValidationError(
+                "При изменении сведений о пользователе произошла ошибка."
+            ) from e
